Migrate additional existing tests to utils/test_coord_transforms.py
